Implement support for multiple (custom) translations of pypy.
The build process is now handled by a generated Makefile and does support
MAKE_JOBS, however since that brings most machines to their knees it is not
declared as MAKE_JOBS_SAFE.

All translations are suffixed with the version and a link is made from the
unsuffixed name to the "default" version at install time via a post-install
script.

Also fix linking of sandbox pypy (missing -L declarations).  Better fix is
available upstream.