Create a `pkg-plist' target to automatically generate the package list.
Move the byte-compiling of the .py files to a post-build stage.  This allows
pkg-plist to properly determine which .py files have a corresponding .pyc (not
all .py files have a .pyc if the .py has a syntax error, or other errors).

Also, cleanup the setting of PYPY_ARCH and PY.  PYPY_ARCH shouldn't have quotes
around the string and do not attempt to set PY if it is already set.  This will
prevent python being pulled in as a dependency if the user knows better.