If PYPY_PRIMARY is not `pypy' then try find pypy if PYPY_PRIMARY does not exist.
PYPY_PRIMARY may not be `pypy' if, for example, PYPY_INST=CLI is used.  In this
case it still may be possible for `pypy' to be installed and it should be used
over cpython.

An alternative is to be more aggressive in finding any pypy binary (i.e.
`ls $LOCALBASE/bin/pypy* | head -n 1`) however there is even less guarantee that
a usable (and suitable) binary could be found.